requirements.txt
pycrypto>=2.6.1
pycryptodome>=3.9.9
termcolor
